Add files via upload
This game is such that you are given a word to spell out and for every incorrect letter, the hangman starts hanging gradually till the life is finally over, then the hangman hangs completely and game is over. However if you correctly spell the word before the life exhausts, then you win.